Consolidate misc components into global
