+++ src/components/global/InProgressLogo/InProgressLogo.module.scss
