+++ src/components/global/InProgressLogo/index.tsx
