+++ src/components/global/LogoToAnimate.tsx
