--- src/components/misc/index.ts
@@ -1,4 +1,0 @@
-import LogoToAnimate from './LogoToAnimate';
-import InProgressLogo from './InProgressLogo';
-
-export { LogoToAnimate, InProgressLogo };
